Move functions from AddTokens to helper file + add comments
